feature/delete_company_confirmation

+++ src/app/components/companies/companies.component.ts
@@ -1,6 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { CompaniesService } from '../../services/companies.service';
 import { ICompany } from '../../interfaces/company.model';
+import { MatDialog } from '@angular/material/dialog';
+import { ConfirmationComponent } from '../dialogs/confirmation/confirmation.component';
 
 @Component({
   selector: 'app-companies',
@@ -10,7 +12,7 @@ import { ICompany } from '../../interfaces/company.model';
 export class CompaniesComponent implements OnInit {
   allCompanies: ICompany[];
 
-  constructor(private companiesService: CompaniesService) { }
+  constructor(private companiesService: CompaniesService, private dialog: MatDialog) { }
 
   ngOnInit(): void {
     this.getCompanies();
@@ -24,9 +26,19 @@ export class CompaniesComponent implements OnInit {
       .subscribe((response) => (this.allCompanies = response));
   }
 
-  deleteCompany(companyId: number): void {
-    this.companiesService.deleteCompany(companyId)
-      .subscribe(() => this.companiesService.reloadCompanies());
+  deleteCompany(companyId: number, companyName: string): void {
+    const confirmDialog = this.dialog.open(ConfirmationComponent, {
+      data: {
+        title: 'Remove Company',
+        message: `Are you sure, you want to delete a company: ${companyName}?`
+      }
+    })
+    confirmDialog.afterClosed().subscribe(result => {
+      if (result === true) {
+        this.companiesService.deleteCompany(companyId)
+          .subscribe(() => this.companiesService.reloadCompanies());
+      }
+    });
   }
 
   searchCompanies(value): void {
